places_lead_scraper/requirements.txt
# Core FastAPI & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.8.2
pydantic-settings==2.2.1
python-dotenv==1.0.0

# Database & ORM
sqlalchemy==2.0.46
psycopg2-binary==2.9.10
alembic==1.13.0
motor==3.3.2

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
PyJWT==2.10.1
cryptography==41.0.7

# API & HTTP
requests==2.31.0
httpx==0.25.2

# Email Validation
email-validator==2.1.0

# Data Processing
pandas==2.2.3
openpyxl==3.1.5
beautifulsoup4==4.12.2

# Caching & Sessions
redis==5.0.1

# Email
python-multipart==0.0.6

# Stripe Payment Processing (optional)
stripe==7.4.0

# Razorpay Payment Processing (optional)
razorpay==1.3.0

# Job Scheduling (optional)
APScheduler==3.10.4

# Logging & Monitoring
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
